Iteration 3 merge to main

+++ app/Controller/routes.py
@@ -1,15 +1,87 @@
+from operator import pos
+from collections import defaultdict
 from flask import Blueprint
 from flask import render_template, flash, redirect, url_for, request
 from config import Config
-from app.Model.models import Position,User,Apply
-from app.Controller.forms import FacultyEditProfileForm, ResearchPositionForm, StudentFilterForm, StudentEditProfileForm, FacultyFilterForm
+from app.Model.models import Position,User, Apply, ClosedPosition
+from app.Controller.forms import ApplicationForm, FacultyEditProfileForm, ResearchPositionForm, StudentFilterForm, StudentEditProfileForm, FacultyFilterForm
 from flask_login import current_user, login_required
 from app import db
+from Constant import researchtopics
 
 
 bp_routes = Blueprint('routes', __name__)
 bp_routes.template_folder = Config.TEMPLATE_FOLDER #'..\\View\\templates'
 
+
+#--------------------------------- Helper Method -----------------------------------------#
+
+
+'''
+@param <student user> <all positions from db>
+@return <ordered list of position from high matched research areas to low matched research areas>
+'''
+def recommandation(student, allpositions):
+    # all the position {position:set(topics)} stored in a dictionary.
+    postion_topic_pair = defaultdict(set)
+    for position in allpositions:
+        p_topics = set()
+        for topic in position.positiontopics:
+            p_topics.add(topic.title)
+        postion_topic_pair[position] = p_topics
+    
+    # student's research topics in a list
+    s_topics = list()
+    for topic in student.researchtopic:
+        s_topics.append(topic.title)
+    
+    # {position:recomm_level}
+    result = {}
+    for pos in postion_topic_pair.keys():
+        recomm_level = 0
+        for s_t in s_topics:
+            if s_t in postion_topic_pair[pos]:
+                recomm_level += 1
+        if recomm_level > 0:
+            result[pos] = recomm_level
+            
+    # rank the positions by recomm level
+    if len(result) > 0:
+        order_pos_level = sorted(list(result.items()), key = lambda x:x[1], reverse=True)
+        order_pos = list(map(lambda x: x[0], order_pos_level))
+        return order_pos
+    # if no such match found then none post show.
+    return []
+
+'''
+Filter by corresponding research topics.
+'''
+def filter_by(sortway, allpositions):
+    if sortway == 'Please choose below options:':
+        return allpositions
+    res = set()  
+    for position in allpositions:
+        for topic in position.positiontopics:
+            if topic.title == sortway: # one position might in multi topic use a set to avoid duplicates
+                res.add(position) 
+    return list(res)
+
+def closed_pos(pos, application):
+    closepos = ClosedPosition(
+        title = pos.title,
+        desc = pos.desc,
+        start_date = pos.start_date,
+        end_date = pos.end_date,
+        time_commitment = pos.time_commitment,
+        applicant_qualification = pos.applicant_qualification,
+        student_id = application.studentid,
+        appstatus = application.status
+    )
+        
+    db.session.add(closepos)
+    db.session.commit()
+    
+#====================================================================================#
 
 #------------------------- Faculty Interfaces ---------------------------------------#
 '''
@@ -18,12 +90,16 @@ Faculty Home Page Route
 @bp_routes.route('/faculty_index', methods=['GET','POST'])
 @login_required
 def faculty_index():
-    positions = Position.query.order_by(Position.time_commitment.desc())
+    if current_user.is_student():
+        return render_template('404error.html', user = current_user)
+    positions = Position.query.order_by(Position.time_commitment.desc()).all()
     fForm = FacultyFilterForm()
     onlymy = False
     if fForm.validate_on_submit():
         onlymy = fForm.checkbox.data
-    return render_template('f_index.html', title="WSU Research Network",positions=positions.all(), onlymy = onlymy, faculty = current_user, form = fForm)
+        sortway = fForm.filter.data
+        positions = filter_by(sortway, positions)
+    return render_template('f_index.html', title="WSU Research Network",positions=positions, onlymy = onlymy, faculty = current_user, form = fForm)
 
 
 '''
@@ -41,10 +117,10 @@ def postReasearch():
                            start_date=postform.start_date.data,
                            end_date=postform.end_date.data, 
                            time_commitment=postform.time_commitment.data, 
-                           research_field=postform.research_field.data, 
                            applicant_qualification=postform.applicant_qualification.data,
                            user_id=current_user.id)
-        
+        for topic in postform.research_field.data:
+            newPost.positiontopics.append(topic)
         db.session.add(newPost)
         db.session.commit()
         flash("Your research position:  " + newPost.title + " is created! ")
@@ -59,8 +135,17 @@ Faculty's delete position route.
 @bp_routes.route('/delete/<position_id>', methods=['POST','DELETE'])
 @login_required
 def delete(position_id):
+    if current_user.is_student():
+        return render_template('404error.html', user = current_user)
     thePost = Position.query.filter_by(id=position_id).first()
     if thePost:
+        for app in thePost.roster:
+            closed_pos(thePost, app)
+            db.session.delete(app)
+            db.session.commit()
+        for r in thePost.positiontopics:
+            thePost.positiontopics.remove(r)
+            db.session.commit()
         db.session.delete(thePost)
         db.session.commit()
         flash("Your Research Position:  " + thePost.title + " has been deleted! ")
@@ -73,9 +158,11 @@ Faculty see all applicants for a paticular position
 @bp_routes.route('/applicants/<position_id>', methods = ['GET'])
 @login_required
 def applicants(position_id):
-    thePost = Position.query.filter_by(id=position_id).first()
+    if current_user.is_student():
+        return render_template('404error.html', user = current_user)
+    thePost = Position.query.filter_by(id=position_id).first() 
     title_str = "Applicants for " + thePost.title
-    return render_template('applicants.html', title = title_str , current_position = thePost)
+    return render_template('applicants.html', title = title_str , current_position = thePost) 
 
 
 '''
@@ -84,9 +171,59 @@ Faculty review applicant list for all position posted
 @bp_routes.route('/applicants_list', methods = ['GET'])
 @login_required
 def applicants_list():
+    if current_user.is_student():
+        return render_template('404error.html', user = current_user)
     position = current_user.get_faculty_posts()
     return render_template('f_applicant_list.html', title = 'Applicant List', pform = position)
 
+'''
+Faculty approve student application for interview
+'''
+@bp_routes.route('/approve/<position_id>/<student_id>', methods=["POST","GET"])
+@login_required
+def approve(position_id,student_id):
+    if current_user.is_student():
+        return render_template('404error.html', user = current_user)
+    application = Apply.query.filter_by(positionid=position_id, studentid = student_id).first()
+    if application:
+        application.status = 2
+        db.session.commit()
+        flash("Student approved for interview!")
+        return redirect(url_for('routes.applicants_list'))
+    return redirect(url_for('routes.applicants_list'))
+
+'''
+Faculty decide to hire student
+'''
+@bp_routes.route('/hire/<position_id>/<student_id>', methods=["POST","GET"])
+@login_required
+def hire(position_id, student_id):
+    if current_user.is_student():
+        return render_template('404error.html', user = current_user)
+    application = Apply.query.filter_by(positionid=position_id, studentid = student_id).first()
+    if application:
+        application.status = 3
+        db.session.commit()
+        flash("Student Hire Successfully!")
+        return redirect(url_for('routes.applicants_list'))
+    return redirect(url_for('routes.applicants_list'))
+
+
+'''
+Faculty decide to reject student
+'''
+@bp_routes.route('/reject/<position_id>/<student_id>', methods=["POST","GET"])
+@login_required
+def reject(position_id, student_id):
+    if current_user.is_student():
+        return render_template('404error.html', user = current_user)
+    application = Apply.query.filter_by(positionid=position_id, studentid = student_id).first()
+    if application:
+        application.status = 4
+        db.session.commit()
+        flash("Student Rejected!")
+        return redirect(url_for('routes.applicants_list'))
+    return redirect(url_for('routes.applicants_list'))
 
 '''
 Faculty display profile
@@ -94,6 +231,8 @@ Faculty display profile
 @bp_routes.route('/f_profile/', methods = ['GET'])
 @login_required
 def f_profile():
+    if current_user.is_student():
+        return render_template('404error.html', user = current_user)
     return render_template('f_profile.html', title='Faculty Profile', faculty = current_user)
 
 '''
@@ -102,6 +241,8 @@ Faculty edit profile
 @bp_routes.route('/f_profile_edit', methods = ['GET','POST'])
 @login_required
 def f_profile_edit():
+    if current_user.is_student():
+        return render_template('404error.html', user = current_user)
     eform = FacultyEditProfileForm()
     if request.method == 'POST':
         #handle the form submission
@@ -127,27 +268,101 @@ def f_profile_edit():
         pass
     return render_template('f_profile_edit.html', title = 'Edit Faculty Profile', form = eform)
 
+'''
+Faculty get student profile
+'''
 @bp_routes.route('/get_s_profile/<s_id>', methods=['GET'])
 @login_required
 def get_s_profile(s_id):
+    if current_user.is_student():
+        return render_template('404error.html', user = current_user)
     theStudent = User.query.filter_by(id=s_id).first()
     titles = theStudent.firstname + ', ' + theStudent.lastname + " Profile"
     return render_template('s_profile.html', title = titles, student = theStudent)
+
+
+'''
+Faculty modify existing position info
+'''
+@bp_routes.route('/f_modify_position<position_id>', methods = ['GET','POST'])
+@login_required
+def f_modify_position(position_id):
+    if current_user.is_student():
+        return render_template('404error.html', user = current_user)
+    thePosition = Position.query.filter_by(id=position_id).first()
+    pform = ResearchPositionForm()
+    if request.method == 'POST':
+        if pform .validate_on_submit():
+            thePosition.title = pform.research_title.data
+            thePosition.desc = pform.desc.data
+            thePosition.start_date = pform.start_date.data
+            thePosition.end_date = pform.end_date.data
+            thePosition.time_commitment = pform.time_commitment.data
+            thePosition.applicant_qualification = pform.applicant_qualification.data
+            for r in thePosition.positiontopics:
+                thePosition.positiontopics.remove(r)
+            for r in pform.research_field.data:
+                thePosition.positiontopics.append(r)
+            db.session.add(thePosition)
+            db.session.commit()
+            flash("Your Position Has Been Updated!")
+            return redirect(url_for('routes.faculty_index'))
+        pass
+    elif request.method == 'GET':
+        pform.research_title.data = thePosition.title
+        pform.desc.data = thePosition.desc
+        pform.start_date.data = thePosition.start_date
+        pform.end_date.data = thePosition.end_date
+        pform.time_commitment.data = thePosition.time_commitment
+        pform.applicant_qualification.data = thePosition.applicant_qualification
+        pform.research_field.data = thePosition.positiontopics
+    else:
+        pass
+    return render_template('f_modify_position.html', position = thePosition, form = pform)
+
+@bp_routes.route('/view_submission/<position_id>/<student_id>', methods = ['GET'])
+@login_required
+def view_submission(position_id,student_id):
+    if current_user.is_student():
+        return render_template('404error.html', user = current_user)
+    position = Position.query.filter_by(id=position_id).first()
+    application = Apply.query.filter_by(positionid=position_id, studentid = student_id).first()
+    return render_template('submission_info.html', application = application, position = position)
 # ==================================================================================#
 
 
+#------------------------------------ Use for both accounts -------------------------------
+@bp_routes.route('/get_position_info/<position_id>', methods=['GET'])
+@login_required
+def get_position_info(position_id):
+    thePosition = Position.query.filter_by(id=position_id).first()
+    application = None
+    if current_user.usertype == 0:
+        application = Apply.query.filter_by(positionid=thePosition.id, studentid = current_user.id).first()
+    return render_template('position_info.html', position = thePosition, application = application)
+
+# ==================================================================================#
+
 # ----------------------------------- Student Interface ----------------------------#
+
 '''
 Student Home Page Route
 '''
 @bp_routes.route('/student_index', methods=['GET','POST'])
 @login_required
 def student_index():
+    if current_user.is_faculty():
+        return render_template('404error.html', user = current_user)
     fForm = StudentFilterForm()
     positions = Position.query.order_by(Position.time_commitment.desc()).all()
     onlyapply = False
     if fForm.validate_on_submit():
         onlyapply = fForm.checkbox.data
+        sortway = fForm.filter.data
+        if sortway == "Recommended Research Opportunities":
+            positions = recommandation(current_user, positions)
+        else:
+            positions = filter_by(sortway, positions)
     return render_template('s_index.html', title = "WSU Research Network", positions=positions, form=fForm, student = current_user, onlyapply = onlyapply)
 
 '''
@@ -156,12 +371,19 @@ Student Apply Position route.
 @bp_routes.route('/apply/<position_id>', methods=['POST'])
 @login_required
 def apply(position_id):
+    if current_user.is_faculty():
+        return render_template('404error.html', user = current_user)
     thePost = Position.query.filter_by(id=position_id).first()
-    if thePost:
-        current_user.apply(thePost)
+    aform = ApplicationForm()
+    if aform.validate_on_submit():
+        fullname = aform.fullname.data 
+        contact_email = aform.email.data 
+        statement = aform.statement.data
+        current_user.apply(thePost, fullname, contact_email, statement)
         db.session.commit()
         flash('You have applied ' + thePost.title +' !')
         return redirect(url_for('routes.student_index'))
+    return render_template('s_application.html', form = aform, position = thePost)
 
 '''
 Student withdraw application route.
@@ -169,17 +391,23 @@ Student withdraw application route.
 @bp_routes.route('/withdraw/<position_id>', methods=['POST'])
 @login_required
 def withdraw(position_id):
+    if current_user.is_faculty():
+        return render_template('404error.html', user = current_user)
     thePost = Position.query.filter_by(id=position_id).first()
     if thePost:
         current_user.withdraw(thePost)
         db.session.commit()
         flash('You have withdraw from the ' + thePost.title +' !')
         return redirect(url_for('routes.student_index'))
+    #TODO: if position is not exist, return error
+    return redirect(url_for('routes.student_index'))
 
 @bp_routes.route('/MyProfile/',methods = ['GET'])
 @login_required
 def My_Profile():
-        return render_template('s_profile.html', title = 'My Profile', student = current_user)
+    if current_user.is_faculty():
+        return render_template('404error.html', user = current_user)
+    return render_template('s_profile.html', title = 'My Profile', student = current_user)
 
 
 '''
@@ -188,6 +416,8 @@ Student edit profile
 @bp_routes.route('/s_profile_edit', methods = ['GET','POST'])
 @login_required
 def s_profile_edit():
+    if current_user.is_faculty():
+        return render_template('404error.html', user = current_user)
     eform = StudentEditProfileForm()
     print(current_user)
     if request.method == 'POST':
